Move compile logic back to mix.exs
It turned out that when it was separated out into a .ex file, the
corresponding .beam file was being included in the release. It seems
really wrong to include a .beam file in the release whose only purpose
is to work around the lack of Makefile support in mix. It will be nice
when Makefile support is added.